Address review feedback for run bundling
