+++ install.sh
@@ -170,21 +170,36 @@ companion_modules=(
     "run-subagent-legacy.mjs"
 )
 
+install_companion_modules() {
+    local source_dir="$1"
+    local installed_count=0
+    local missing=()
+
+    for module in "${companion_modules[@]}"; do
+        if [[ -f "$source_dir/$module" ]]; then
+            mv "$source_dir/$module" "$COMPOSIO_INSTALL_DIR/$module"
+            installed_count=$((installed_count + 1))
+        else
+            missing+=("$module")
+        fi
+    done
+
+    if (( installed_count > 0 && installed_count < ${#companion_modules[@]} )); then
+        error "Downloaded archive is incomplete; missing companion modules: ${missing[*]}"
+    fi
+
+    if (( installed_count == 0 )); then
+        warn "This release archive does not include the companion modules required by 'composio run'. That command may be unavailable in this version."
+    fi
+}
+
 # Handle nested directory structure (composio-<target>/composio)
 if [[ -f "$tmpdir/composio-$target/composio" ]]; then
     mv "$tmpdir/composio-$target/composio" "$exe"
-    for module in "${companion_modules[@]}"; do
-        if [[ -f "$tmpdir/composio-$target/$module" ]]; then
-            mv "$tmpdir/composio-$target/$module" "$COMPOSIO_INSTALL_DIR/$module"
-        fi
-    done
+    install_companion_modules "$tmpdir/composio-$target"
 elif [[ -f "$tmpdir/composio" ]]; then
     mv "$tmpdir/composio" "$exe"
-    for module in "${companion_modules[@]}"; do
-        if [[ -f "$tmpdir/$module" ]]; then
-            mv "$tmpdir/$module" "$COMPOSIO_INSTALL_DIR/$module"
-        fi
-    done
+    install_companion_modules "$tmpdir"
 else
     error 'Binary not found in extracted archive'
 fi
