+++ ts/packages/cli/scripts/build-binary-cross.ts
@@ -15,12 +15,10 @@
  */
 
 import process from 'node:process';
-import { mkdir } from 'node:fs/promises';
 import { Config, ConfigProvider, Console, Effect, Stream, Logger, Layer, LogLevel } from 'effect';
 import { Command } from '@effect/platform';
 import { BunContext, BunRuntime } from '@effect/platform-bun';
-import { RUN_COMPANION_MODULE_BASENAMES } from '../src/services/run-companion-modules';
-import { teardown } from './_shared';
+import { buildCompanionModules, teardown } from './_shared';
 
 /**
  * Maps Bun cross-compilation targets to Composio artifact names.
@@ -109,47 +107,7 @@ export function buildBinaryCross() {
     }
 
     const companionOutputDir = './dist/binaries/companions';
-    yield* Effect.tryPromise(() => mkdir(companionOutputDir, { recursive: true }));
-    for (const name of RUN_COMPANION_MODULE_BASENAMES) {
-      const companionArgs = [
-        'bun',
-        'build',
-        `./src/services/${name}.ts`,
-        '--outfile',
-        `${companionOutputDir}/${name}.mjs`,
-        '--format',
-        'esm',
-        '--target',
-        'bun',
-      ] as const satisfies ReadonlyArray<string>;
-
-      const companionCmd = Command.make(...companionArgs);
-      const { exitCode: companionExitCode } = yield* companionCmd.pipe(
-        Command.start,
-        Effect.flatMap(process =>
-          Effect.all(
-            {
-              exitCode: process.exitCode,
-              output: Stream.merge(
-                Stream.decodeText(process.stdout, 'utf-8'),
-                Stream.decodeText(process.stderr, 'utf-8'),
-                { haltStrategy: 'left' }
-              ).pipe(
-                Stream.tap(chunk => Console.log(chunk)),
-                Stream.runDrain
-              ),
-            },
-            {
-              concurrency: 'unbounded',
-            }
-          )
-        )
-      );
-
-      if (companionExitCode !== 0) {
-        return yield* Effect.fail(new Error(`Failed to build companion module: ${name}`));
-      }
-    }
+    yield* buildCompanionModules(companionOutputDir);
 
     yield* Console.log(`Binary cross-compiled: ${outfile}`);
   });
